requirements.txt
streamlit
pdfplumber
pytesseract
Pillow
openai
pandas
requests
beautifulsoup4
